Add the SCR agent for /etc/vconsole.conf
It was removed as part of the keyboard refactorization recently done
(see 4e32e3a), but still necessary for the yast-country/console module
to set the console font since systemd is not able to manage it at this
time.

So, the agent has been "restored" but placing it in the
yast-country/console instead in yast-country/keyboard.

Just for the record, it was originally introduced in https://github.com/yast/yast-country/pull/30,
more specifically in cd0e0dae5f99.