Cleanup: use "ru" keymap for Russian, not "ruwin_alt-UTF-8" (bsc#1194609)
ruwin_alt-UTF-8 was introduced in 2008 in response to
https://bugzilla.suse.com/show_bug.cgi?id=432862
because the previous keymap did not use Unicode.

"ru" does use Unicode, and it has a dual latin-cyrillic layout
(AltGr for single characters, Ctrl-Shift for layout switching)

In /usr/share/systemd/kbd-model-map the Russian keymap is a
kbd-model-map.legacy SUSE patch over upstream, and as Franck Bui was
reviewing this patch for upstreaming, we found that an entry for Russian
is still needed, but a default "ru" keymap has a better chance than a
specific one.